pyproject.toml
[project]
name = "algokit-templates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "copier>=9.5.0",
  "fire>=0.7.0",
  "pydantic>=2.10.6",
  "pyyaml>=6.0.2",
  "tomli>=2.2.1",
  "tomli-w>=1.2.0",
]

[tool.uv]
dev-dependencies = [
  "mypy>=1.16.1",
  "pytest-sugar>=1.0.0",
  "pytest>=8.4.0",
  "ruff>=0.12.0",
  "types-pyyaml>=6.0.12.20250516",
]
